pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "guitar-practice-assistant"
version = "0.1.0"
description = "A guitar practice routine manager"
authors = [
    {name = "Steven"}
]
requires-python = ">=3.10"
dependencies = [
    "flask",
    "python-dotenv",
    "google-auth",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "gspread",
    "anthropic",
    "python-magic",
    "pytesseract",
    "pdf2image",
    "pillow",
    "sqlalchemy",
    "psycopg2-binary",
    "flask-cors"
]

[tool.setuptools]
packages = ["app"]

[tool.pylance]
python.analysis.extraPaths = [".", "app"]